Add files via upload
⚠️Disclaimer⚠️

This login screen design is solely created for educational and demonstration purposes. It is not associated with Pinterest or any affiliated entities. The use of Pinterest's name, logo, or any other trademarked materials within this project is purely for illustrative purposes. This project was developed independently as a practice exercise to showcase HTML and CSS skills. Any resemblance to existing products, services, or designs is coincidental. The creator of this login screen is not responsible for any misuse, misinterpretation, or unauthorized usage of this design.